Pre-release commit

+++ src/main/java/services/ExaminationService.java
@@ -1,5 +1,6 @@
 package services;
 
+import enums.HospitalizationRelation;
 import exceptions.EntityNotFoundException;
 import exceptions.ErrorMessageKeysContainedException;
 import exceptions.UnknownSqlException;
@@ -7,6 +8,8 @@ import model.dao.implementations.mysql.MySqlDaoFactory;
 import model.dao.interfaces.*;
 import model.database.TransactionManager;
 import model.entities.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import utils.LongLimit;
 import utils.PageContent;
 
@@ -17,12 +20,15 @@ import java.util.stream.Collectors;
 
 public class ExaminationService {
 
+    private static final Logger LOGGER = LoggerFactory.getLogger(ExaminationService.class);
     private static DaoFactory daoFactory = MySqlDaoFactory.getInstance();
     private static ExaminationDao examinationDao = daoFactory.createExaminationDao();
     private static DiagnoseDao diagnoseDao = daoFactory.createDiagnoseDao();
     private static UserDao userDao = daoFactory.createUserDao();
     private static AssignmentDao assignmentDao = daoFactory.createAssignmentDao();
     private static AssignmentTypeDao assignmentTypeDao = daoFactory.createAssignmentTypeDao();
+    private static HospitalizationDao hospitalizationDao = daoFactory.createHospitalizationDao();
+    private static RoleDao roleDao = daoFactory.createRoleDao();
 
     private ExaminationService() {
 
@@ -30,9 +36,25 @@ public class ExaminationService {
 
     public static long addExamination(Examination examination) {
         try {
+            TransactionManager.beginTransaction();
+            Hospitalization hospitalization = new Hospitalization();
+            long patientId = examination.getPatient().getId();
+            long currentPatientHospitalizationId = hospitalizationDao.getIdOfCurrentHospitalizationByPatientId(patientId);
+            if (examination.getHospitalizationRelation() == HospitalizationRelation.INITIAL) {
+                hospitalization.setId(hospitalizationDao.insert(hospitalization));
+                userDao.updateHospitalizedStatus(true, patientId);
+            } else if (examination.getHospitalizationRelation() == HospitalizationRelation.DISCHARGE) {
+                hospitalization.setId(currentPatientHospitalizationId);
+                userDao.updateHospitalizedStatus(false, patientId);
+            } else if (examination.getHospitalizationRelation() == HospitalizationRelation.INTERMEDIATE) {
+                hospitalization.setId(currentPatientHospitalizationId);
+            }
+            examination.setHospitalization(hospitalization);
             long examinationId = examinationDao.insert(examination);
+            TransactionManager.commitTransaction();
             return examinationId;
         } catch (UnknownSqlException e) {
+            TransactionManager.rollbackTransaction();
             e.printStackTrace();
             throw e;
         }
@@ -49,8 +71,16 @@ public class ExaminationService {
 
     public static void deleteExamination(long examinationId) {
         try {
-            examinationDao.delete(examinationId);
+            TransactionManager.beginTransaction();
+            Examination examination = examinationDao.selectById(examinationId);
+            if (examination.getHospitalizationRelation() == HospitalizationRelation.SINGLE ||
+                    examination.getHospitalizationRelation() == HospitalizationRelation.INTERMEDIATE) {
+                examinationDao.delete(examinationId);
+            } else
+                throw new ErrorMessageKeysContainedException(List.of("examination.undeletable"));
+            TransactionManager.commitTransaction();
         } catch (UnknownSqlException e) {
+            TransactionManager.rollbackTransaction();
             e.printStackTrace();
             throw e;
         }
@@ -119,9 +149,27 @@ public class ExaminationService {
         LongLimit longLimit = new LongLimit(offset, itemsPerPage);
         List<Examination> content = examinationDao.selectAllInRange(longLimit);
         content.forEach(ExaminationService::fillExamination);
-        long countOfExaminationsWithDoctorId = examinationDao.selectCountOfExaminations();
-        int totalPages = (int)((countOfExaminationsWithDoctorId / itemsPerPage) +
-                (countOfExaminationsWithDoctorId % itemsPerPage == 0 ? 0 : 1));
+        long countOfExaminations = examinationDao.selectCountOfExaminations();
+        int totalPages = (int)((countOfExaminations / itemsPerPage) +
+                (countOfExaminations % itemsPerPage == 0 ? 0 : 1));
+        PageContent<Examination> examinationPageContent = new PageContent<>();
+        examinationPageContent.setContent(content);
+        examinationPageContent.setPage(page);
+        examinationPageContent.setTotalPages(totalPages);
+        return examinationPageContent;
+    }
+
+    public static PageContent<Examination> getIntermediateExaminationsForPageByHospitalizationId(
+            long hospitalizationId, int page, int itemsPerPage) {
+        long offset = (page - 1) * itemsPerPage;
+        LongLimit longLimit = new LongLimit(offset, itemsPerPage);
+        List<Examination> content = examinationDao.selectIntermediateExaminationsByHospitalizationIdInRange(
+                hospitalizationId, longLimit);
+        content.forEach(ExaminationService::fillExamination);
+        long countOfIntermediateExaminationsWithHospitalizationId =
+                examinationDao.selectCountOfIntermediateExaminationsWithHospitalizationId(hospitalizationId);
+        int totalPages = (int)((countOfIntermediateExaminationsWithHospitalizationId / itemsPerPage) +
+                (countOfIntermediateExaminationsWithHospitalizationId % itemsPerPage == 0 ? 0 : 1));
         PageContent<Examination> examinationPageContent = new PageContent<>();
         examinationPageContent.setContent(content);
         examinationPageContent.setPage(page);
@@ -130,30 +178,39 @@ public class ExaminationService {
     }
 
     public static Map<String, Object> getDropDownsData() {
-        RoleDao roleDao = daoFactory.createRoleDao();
-        AssignmentTypeDao assignmentTypeDao = daoFactory.createAssignmentTypeDao();
-        UserDao userDao = daoFactory.createUserDao();
+        TransactionManager.beginTransaction();
         Map<String, Object> outerMap = new HashMap<>();
-        List<Role> roles = roleDao.selectAll();
-        List<AssignmentType> assignmentTypes = assignmentTypeDao.selectAll();
-        Map<Long, List<Role>> assignmentTypeIdsToRoleMap;
-        Map<Long, List<User>> roleIdsToExecutorMap;
-        assignmentTypeIdsToRoleMap = assignmentTypeDao.selectAll().stream().collect(Collectors.toMap(
-                (AssignmentType assignmentType) -> assignmentType.getId(),
-                (AssignmentType assignmentType) -> roleDao.selectByAssignmentTypeId(assignmentType.getId())));
-        roleIdsToExecutorMap = roleDao.selectAll().stream().collect(Collectors.toMap(
-                (Role role) -> role.getId(),
-                (Role role) -> userDao.selectAllShortByRoleId(role.getId())));
-        outerMap.put("roles", roles);
-        outerMap.put("assignmentTypes", assignmentTypes);
-        outerMap.put("assignmentTypeIdsToRoleMap", assignmentTypeIdsToRoleMap);
-        outerMap.put("roleIdsToExecutorMap", roleIdsToExecutorMap);
+        try {
+
+            List<Role> roles = roleDao.selectAll();
+            List<AssignmentType> assignmentTypes = assignmentTypeDao.selectAll();
+            Map<Long, List<Role>> assignmentTypeIdsToRoleMap;
+            Map<Long, List<User>> roleIdsToExecutorMap;
+            assignmentTypeIdsToRoleMap = assignmentTypeDao.selectAll().stream().collect(Collectors.toMap(
+                    (AssignmentType assignmentType) -> assignmentType.getId(),
+                    (AssignmentType assignmentType) -> roleDao.selectByAssignmentTypeId(assignmentType.getId())));
+            roleIdsToExecutorMap = roleDao.selectAll().stream().collect(Collectors.toMap(
+                    (Role role) -> role.getId(),
+                    (Role role) -> userDao.selectAllShortByRoleIdAndHospitalizationStatus(role.getId(), false)));
+            System.out.println("Map: " + roles);
+            outerMap.put("roles", roles);
+            outerMap.put("assignmentTypes", assignmentTypes);
+            outerMap.put("assignmentTypeIdsToRoleMap", assignmentTypeIdsToRoleMap);
+            outerMap.put("roleIdsToExecutorMap", roleIdsToExecutorMap);
+            TransactionManager.commitTransaction();
+        } catch (UnknownSqlException e) {
+            LOGGER.error("Error fetching drop-down data: ", e);
+            TransactionManager.rollbackTransaction();
+        }
         return outerMap;
     }
 
-    private static void fillExamination(Examination examination) {
+    static void fillExamination(Examination examination) {
         examination.setPatient(userDao.selectById(examination.getPatient().getId()));
         examination.setDoctor(userDao.selectById(examination.getDoctor().getId()));
         examination.setDiagnoses(diagnoseDao.selectByExaminationId(examination.getId()));
+        if (examination.getHospitalization() != null) {
+            examination.setHospitalization(hospitalizationDao.selectById(examination.getHospitalization().getId()));
+        }
     }
 }
